test: Move update_user_credits tests to test_doc2sys_user_settings.py and remove redundant test_payment_integration.py
